https://github.com/metanorma/iso-10303/issues/380
